+++ cmake/thirdparty/BLTSetupCUDA.cmake
@@ -1,28 +1,15 @@
 # Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level LICENSE file for details
-# 
+#
 # SPDX-License-Identifier: (BSD-3-Clause)
-
-# Use legacy FindCUDA. This was deprecated in 3.10 and removed in 3.27.
-# TODO: Do not rely on this setting and use the new CUDA finding mechanism instead:
-# https://cmake.org/cmake/help/latest/module/FindCUDA.html
-if(POLICY CMP0146)
-  cmake_policy(SET CMP0146 OLD)
-endif()
 
 ################################
 # Sanity Checks
 ################################
 
-# Rare case of two flags being incompatible
-if (DEFINED CMAKE_SKIP_BUILD_RPATH AND DEFINED CUDA_LINK_WITH_NVCC)
-    if (NOT CMAKE_SKIP_BUILD_RPATH AND CUDA_LINK_WITH_NVCC)
-        message( FATAL_ERROR
-                 "CMAKE_SKIP_BUILD_RPATH (FALSE) and CUDA_LINK_WITH_NVCC (TRUE) "
-                 "are incompatible when linking explicit shared libraries. Set "
-                 "CMAKE_SKIP_BUILD_RPATH to TRUE.")
-    endif()
-endif()
+if( ${CMAKE_VERSION} VERSION_LESS "3.17.0" )
+  message(FATAL_ERROR "CUDA support requires CMake >= 3.17.0")
+endif ()
 
 # CMAKE_CUDA_HOST_COMPILER needs to be set prior to enabling the CUDA language
 get_property(_languages GLOBAL PROPERTY ENABLED_LANGUAGES)
@@ -42,100 +29,70 @@ if ( NOT CMAKE_CUDA_HOST_COMPILER )
     endif()
 endif()
 
-# Override rpath link flags for nvcc
-if (CUDA_LINK_WITH_NVCC)
-    set(CMAKE_SHARED_LIBRARY_RUNTIME_CUDA_FLAG "-Xlinker -rpath -Xlinker " CACHE STRING "")
-    set(CMAKE_SHARED_LIBRARY_RPATH_LINK_CUDA_FLAG "-Xlinker -rpath -Xlinker " CACHE STRING "")
+if(CUDA_TOOLKIT_ROOT_DIR AND NOT CUDAToolkit_ROOT)
+    set(CUDAToolkit_ROOT "${CUDA_TOOLKIT_ROOT_DIR}" CACHE PATH
+        "Root directory of the CUDA Toolkit" FORCE)
+endif()
+
+if(DEFINED CUDA_SEPARABLE_COMPILATION AND NOT DEFINED CMAKE_CUDA_SEPARABLE_COMPILATION)
+    set(CMAKE_CUDA_SEPARABLE_COMPILATION "${CUDA_SEPARABLE_COMPILATION}" CACHE BOOL
+        "Build CUDA objects with separable compilation enabled" FORCE)
 endif()
 
 
 ############################################################
 # Basics
 ############################################################
-# language check fails when using clang-cuda
-if (NOT BLT_ENABLE_CLANG_CUDA)
-    enable_language(CUDA)
-endif()
-
-############################################################
-# Map Legacy FindCUDA variables to native cmake variables
-############################################################
-# if we are linking with NVCC, define the link rule here
-# Note that some mpi wrappers might have things like -Wl,-rpath defined, which when using 
-# FindMPI can break nvcc. In that case, you should set ENABLE_FIND_MPI to Off and control
-# the link using CMAKE_CUDA_LINK_FLAGS. -Wl,-rpath, equivalent would be -Xlinker -rpath -Xlinker
-if (CUDA_LINK_WITH_NVCC)
-    set(CMAKE_CUDA_LINK_EXECUTABLE
-    "${CMAKE_CUDA_COMPILER} <CMAKE_CUDA_LINK_FLAGS>  <FLAGS>  <LINK_FLAGS>  <OBJECTS> -o <TARGET>  <LINK_LIBRARIES>")
-    # do a no-op for the device links - for some reason the device link library dependencies are only a subset of the 
-    # executable link dependencies so the device link fails if there are any missing CUDA library dependencies. Since
-    # we are doing a link with the nvcc compiler, the device link step is unnecessary .
-    # Frustratingly, nvcc-link errors out if you pass it an empty file, so we have to first compile the empty file. 
-    set(CMAKE_CUDA_DEVICE_LINK_LIBRARY "touch <TARGET>.cu ; ${CMAKE_CUDA_COMPILER} <CMAKE_CUDA_LINK_FLAGS> -std=c++${CMAKE_CUDA_STANDARD} -dc <TARGET>.cu -o <TARGET>")
-    set(CMAKE_CUDA_DEVICE_LINK_EXECUTABLE "touch <TARGET>.cu ; ${CMAKE_CUDA_COMPILER} <CMAKE_CUDA_LINK_FLAGS> -std=c++${CMAKE_CUDA_STANDARD} -dc <TARGET>.cu -o <TARGET>")
-endif()
+enable_language(CUDA)
 
 ############################################################
 # Find CUDA
 ############################################################
-# If CUDA_TOOLKIT_ROOT_DIR is not set, it should be set by find_package(CUDA)
-find_package(CUDA REQUIRED)
-blt_assert_exists( DIRECTORIES ${CUDA_TOOLKIT_ROOT_DIR} )
+find_package(CUDAToolkit REQUIRED)
+blt_assert_exists( TARGETS CUDA::cudart )
 
-############################################################
-# Append the path to the NVIDIA SDK to the link flags
-############################################################
-if ( IS_DIRECTORY "${CUDA_TOOLKIT_ROOT_DIR}/lib64" )
-    list(APPEND CMAKE_CUDA_LINK_FLAGS "-L${CUDA_TOOLKIT_ROOT_DIR}/lib64" )
+if(NOT CUDAToolkit_ROOT AND CUDAToolkit_BIN_DIR)
+    get_filename_component(CUDAToolkit_ROOT "${CUDAToolkit_BIN_DIR}" DIRECTORY)
 endif()
 
-if ( IS_DIRECTORY "${CUDA_TOOLKIT_ROOT_DIR}/lib}" )
-    list(APPEND CMAKE_CUDA_LINK_FLAGS "-L${CUDA_TOOLKIT_ROOT_DIR}/lib" )
-endif()
+# Provide compatibility variables for projects that still query the old names.
+set(CUDA_FOUND ${CUDAToolkit_FOUND})
+set(CUDA_VERSION_STRING ${CUDAToolkit_VERSION})
+set(CUDA_INCLUDE_DIRS ${CUDAToolkit_INCLUDE_DIRS})
+set(CUDA_LIBRARIES CUDA::cudart)
 
 ############################################################
 # Output information about CUDA
 ############################################################
-message(STATUS "CUDA Version:                   ${CUDA_VERSION_STRING}")
-message(STATUS "CUDA Toolkit Root Dir:          ${CUDA_TOOLKIT_ROOT_DIR}")
+message(STATUS "CUDA Version:                   ${CUDAToolkit_VERSION}")
+message(STATUS "CUDA Toolkit Root Dir:          ${CUDAToolkit_ROOT}")
 message(STATUS "CUDA Compiler:                  ${CMAKE_CUDA_COMPILER}")
 message(STATUS "CUDA Host Compiler:             ${CMAKE_CUDA_HOST_COMPILER}")
 message(STATUS "CUDA Standard:                  ${CMAKE_CUDA_STANDARD}")
 if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.18" )
 message(STATUS "CUDA Architectures:             ${CMAKE_CUDA_ARCHITECTURES}")
 endif()
-message(STATUS "CUDA Include Path:              ${CUDA_INCLUDE_DIRS}")
-message(STATUS "CUDA Libraries:                 ${CUDA_LIBRARIES}")
+message(STATUS "CUDA Include Path:              ${CUDAToolkit_INCLUDE_DIRS}")
+message(STATUS "CUDA Runtime Target:            CUDA::cudart")
 message(STATUS "CUDA Compile Flags:             ${CMAKE_CUDA_FLAGS}")
 message(STATUS "CUDA Link Flags:                ${CMAKE_CUDA_LINK_FLAGS}")
-message(STATUS "BLT CUDA Separable Compilation: ${CUDA_SEPARABLE_COMPILATION}")
 message(STATUS "CUDA Separable Compilation:     ${CMAKE_CUDA_SEPARABLE_COMPILATION}")
-message(STATUS "CUDA Link with NVCC:            ${CUDA_LINK_WITH_NVCC}")
 message(STATUS "CUDA Implicit Link Libraries:   ${CMAKE_CUDA_IMPLICIT_LINK_LIBRARIES}")
 message(STATUS "CUDA Implicit Link Directories: ${CMAKE_CUDA_IMPLICIT_LINK_DIRECTORIES}")
 
 ############################################################
 # Check CUDA language standard is supported
 ############################################################
-# Note: change to use CUDAToolkit_VERSION after moving away from find_package(CUDA)
 if(CMAKE_CUDA_STANDARD STREQUAL "17")
     if(NOT DEFINED CMAKE_CUDA_COMPILE_FEATURES OR (NOT "cuda_std_17" IN_LIST CMAKE_CUDA_COMPILE_FEATURES))
-        message(FATAL_ERROR "CUDA ${CUDA_VERSION_STRING} does not support C++17.")
+        message(FATAL_ERROR "CUDA ${CUDAToolkit_VERSION} does not support C++17.")
     endif()
 endif()
 
 if(CMAKE_CUDA_STANDARD STREQUAL "20")
     if(NOT DEFINED CMAKE_CUDA_COMPILE_FEATURES OR (NOT "cuda_std_20" IN_LIST CMAKE_CUDA_COMPILE_FEATURES))
-        message(FATAL_ERROR "CUDA ${CUDA_VERSION_STRING} does not support C++20.")
+        message(FATAL_ERROR "CUDA ${CUDAToolkit_VERSION} does not support C++20.")
     endif()
-endif()
-
-# don't propagate host flags - too easy to break stuff!
-set (CUDA_PROPAGATE_HOST_FLAGS Off)
-if (CMAKE_CXX_COMPILER)
-    set(CUDA_HOST_COMPILER ${CMAKE_CXX_COMPILER})
-else()
-    set(CUDA_HOST_COMPILER ${CMAKE_C_COMPILER})
 endif()
 
 # Set PIE options to empty for PGI since it doesn't understand -fPIE This
@@ -146,11 +103,29 @@ if("${CMAKE_CXX_COMPILER_ID}" STREQUAL "PGI")
 endif()
 
 
-set(_cuda_compile_flags " ")
-if (BLT_ENABLE_CLANG_CUDA)
-    set (_cuda_compile_flags -x cuda --cuda-gpu-arch=${BLT_CLANG_CUDA_ARCH} --cuda-path=${CUDA_TOOLKIT_ROOT_DIR})
-    message(STATUS "Clang CUDA Enabled. CUDA compile flags added: ${_cuda_compile_flags}")    
+# CUDA targets must be global for aliases when created as imported targets.
+set(_blt_cuda_is_global On)
+if(${BLT_EXPORT_THIRDPARTY})
+    set(_blt_cuda_is_global Off)
 endif()
+
+# Use DEPENDS_ON to keep the CUDA::cudart imported library target in the link
+# interface. blt_inherit_target_info() only copies interface usage properties.
+
+# Use the CUDA runtime without flagging source files as
+# CUDA language.  This causes your source files to use
+# the regular C/CXX compiler. This is separate from
+# linking with nvcc.
+# This logic is handled in the blt_add_library/executable
+# macros
+blt_import_library(NAME          cuda_runtime
+                   INCLUDES      ${CUDAToolkit_INCLUDE_DIRS}
+                   TREAT_INCLUDES_AS_SYSTEM ON
+                   DEPENDS_ON    CUDA::cudart
+                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY}
+                   GLOBAL        ${_blt_cuda_is_global})
+
+add_library(blt::cuda_runtime ALIAS cuda_runtime)
 
 # depend on 'cuda', if you need to use cuda
 # headers, link to cuda libs, and need to compile your
@@ -159,23 +134,8 @@ endif()
 # This logic is handled in the blt_add_library/executable
 # macros
 blt_import_library(NAME          cuda
-                   COMPILE_FLAGS ${_cuda_compile_flags}
-                   INCLUDES      ${CUDA_INCLUDE_DIRS}
-                   LIBRARIES     ${CUDA_LIBRARIES}
-                   LINK_FLAGS    "${CMAKE_CUDA_LINK_FLAGS}"
-                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})
+                   DEPENDS_ON    cuda_runtime
+                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY}
+                   GLOBAL        ${_blt_cuda_is_global})
 
 add_library(blt::cuda ALIAS cuda)
-
-# same as 'cuda' but we don't flag your source files as
-# CUDA language.  This causes your source files to use 
-# the regular C/CXX compiler. This is separate from 
-# linking with nvcc.
-# This logic is handled in the blt_add_library/executable
-# macros
-blt_import_library(NAME       cuda_runtime
-                   INCLUDES   ${CUDA_INCLUDE_DIRS}
-                   LIBRARIES  ${CUDA_LIBRARIES}
-                   EXPORTABLE ${BLT_EXPORT_THIRDPARTY})
-
-add_library(blt::cuda_runtime ALIAS cuda_runtime)